+++ pkg/readings/models.go
@@ -2,5 +2,5 @@ package readings
 
 type Reading struct {
 	SensorId string  `json:"sensorId"`
-	Moisture float32 `json:"moisture"`
+	Moisture float64 `json:"moisture"`
 }
